move mainMenu to ui
