+++ src/ts/cosmosJourneyer.ts
@@ -31,7 +31,7 @@ import "@babylonjs/core/Engines/WebGPU/Extensions/";
 import { PauseMenu } from "./ui/pauseMenu";
 import { StarSystemView } from "./starSystem/starSystemView";
 import { EngineFactory } from "@babylonjs/core/Engines/engineFactory";
-import { MainMenu } from "./mainMenu/mainMenu";
+import { MainMenu } from "./ui/mainMenu";
 import { SystemSeed } from "./utils/systemSeed";
 import { SaveFileData } from "./saveFile/saveFileData";
 import { Vector3 } from "@babylonjs/core/Maths/math.vector";
